pyproject.toml
[project]
name = "crypto-market"
version = "1.0.0"
description = "Crypto Market where you can see all the Coins and categories etc.."
authors = [
    {name = "shyamkumar.ss724@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.1.5,<6.0.0)",
    "django-extensions (>=3.2.3,<4.0.0)",
    "django-filter (>=24.3,<25.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "flake8 (>=7.1.1,<8.0.0)",
    "black (>=24.10.0,<25.0.0)",
    "pre-commit (>=4.1.0,<5.0.0)",
    "django-dotenv (>=1.4.2,<2.0.0)",
    "drf-spectacular-sidecar (>=2024.12.1,<2025.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
